+++ packages/app-builder/src/services/ast-node/getAstNodeDisplayName.ts
@@ -19,7 +19,7 @@ import {
   type TimestampExtractAstNode,
 } from '@app-builder/models/ast-node';
 import { type CustomList } from '@app-builder/models/custom-list';
-import { getOperatorName } from '@app-builder/models/getOperatorName';
+import { getOperatorName } from '@app-builder/models/get-operator-name';
 import { isAggregatorOperator } from '@app-builder/models/modale-operators';
 import { formatNumber } from '@app-builder/utils/format';
 import { type TFunction } from 'i18next';
